# docs/swagger.yaml
definitions:
...
  database.Dataset:
    properties:
      created_at:
        type: string
      id:
        type: integer
      repository:
        $ref: '#/definitions/database.Repository'
      repository_id:
        type: integer
      updated_at:
        type: string
    type: object
  database.Mirror:
    properties:
      created_at:
        type: string
      id:
        type: integer
      interval:
        type: string
      last_message:
        type: string
      last_updated_at:
        type: string
      local_repo_path:
        type: string
      mirror_source:
        $ref: '#/definitions/database.MirrorSource'
      mirror_source_id:
        type: integer
      mirror_task_id:
        type: integer
      next_execution_timestamp:
        type: string
      priority:
        $ref: '#/definitions/types.MirrorPriority'
      progress:
        type: integer
      push_mirror_created:
        type: boolean
      repository:
        $ref: '#/definitions/database.Repository'
      repository_id:
        type: integer
      source_repo_path:
        type: string
      source_url:
        type: string
      status:
        $ref: '#/definitions/types.MirrorTaskStatus'
      updated_at:
        type: string
    type: object
...
  database.Repository:
    properties:
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      download_count:
        type: integer
      downloads:
        items:
          $ref: '#/definitions/database.RepositoryDownload'
        type: array
      git_path:
        type: string
      http_clone_url:
        type: string
      id:
        type: integer
      labels:
        description: Depreated
        type: string
      license:
        type: string
      likes:
        type: integer
      mirror:
        $ref: '#/definitions/database.Mirror'
      name:
        type: string
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      readme:
        description: Depreated
        type: string
      repository_type:
        $ref: '#/definitions/types.RepositoryType'
      sensitive_check_status:
        $ref: '#/definitions/types.SensitiveCheckStatus'
      source:
        $ref: '#/definitions/types.RepositorySource'
      ssh_clone_url:
        type: string
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/database.Tag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/database.User'
      user_id:
        type: integer
    type: object
  database.RepositoryDownload:
    properties:
      click_download_count:
        type: integer
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      repository:
        $ref: '#/definitions/database.Repository'
      repository_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  types.APIBadRequest:
    type: object
  types.APIInternalServerError:
    type: object
...
  types.Code:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      downloads:
        type: integer
      id:
        type: integer
      license:
        type: string
      likes:
        type: integer
      name:
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      source:
        $ref: '#/definitions/types.RepositorySource'
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
    type: object
...
  types.Commit:
    properties:
      author_email:
        type: string
      author_name:
        type: string
      authored_date:
        type: string
      committer_date:
        type: string
      committer_email:
        type: string
      committer_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
    type: object
...
  types.CreateDatasetReq:
    properties:
      default_branch:
        example: main
        type: string
      description:
        type: string
      labels:
        example: ""
        type: string
      license:
        example: MIT
        type: string
      name:
        example: model_name_1
        type: string
      namespace:
        example: user_or_org_name
        type: string
      nickname:
        example: model display name
        type: string
      private:
        type: boolean
      readme:
        type: string
      type:
        type: integer
    type: object
...
  types.Dataset:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      downloads:
        type: integer
      id:
        type: integer
      license:
        type: string
      likes:
        type: integer
      name:
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      readme:
        type: string
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      source:
        $ref: '#/definitions/types.RepositorySource'
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
    type: object
...
  types.File:
    properties:
      commit:
        $ref: '#/definitions/types.Commit'
      content:
        type: string
      last_commit_sha:
        type: string
      lfs:
        type: boolean
      lfs_pointer_size:
        type: integer
      lfs_relative_path:
        description: relative path in lfs storage
        type: string
      mode:
        type: string
      name:
        type: string
      path:
        type: string
      sha:
        type: string
      size:
        type: integer
      type:
        type: string
      url:
        description: URL to browse the file
        type: string
    type: object
...
  types.Model:
    properties:
      base_model:
        type: string
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      downloads:
        type: integer
      enable_finetune:
        type: boolean
      enable_inference:
        type: boolean
      id:
        type: integer
      license:
        type: string
      likes:
        type: integer
      name:
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      readme:
        type: string
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      source:
        $ref: '#/definitions/types.RepositorySource'
      status:
        description: url to interact with the model
        example: RUNNING
        type: string
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
      widget_type:
        allOf:
        - $ref: '#/definitions/types.ModelWidgetType'
        description: 'widget UI style: generation,chat'
        example: generation
    type: object
...
  types.Namespace:
    properties:
      avatar:
        type: string
      path:
        type: string
      type:
        description: namespace types like 'user' for normal user, and 'school', 'company'
          for orgs etc.
        type: string
    type: object
  types.Organization:
    properties:
      homepage:
        type: string
      logo:
        type: string
      name:
        type: string
      org_type:
        type: string
      path:
        description: unique name of the organization
        type: string
      user_id:
        type: integer
      verified:
        type: boolean
    type: object
...
  types.PromptRes:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      downloads:
        type: integer
      id:
        type: integer
      license:
        type: string
      likes:
        type: integer
      name:
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      readme:
        type: string
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      source:
        $ref: '#/definitions/types.RepositorySource'
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
    type: object
...
  types.Relations:
    properties:
      codes:
        items:
          $ref: '#/definitions/types.Code'
        type: array
      datasets:
        items:
          $ref: '#/definitions/types.Dataset'
        type: array
      models:
        items:
          $ref: '#/definitions/types.Model'
        type: array
      prompts:
        items:
          $ref: '#/definitions/types.PromptRes'
        type: array
      spaces:
        items:
          $ref: '#/definitions/types.Space'
        type: array
    type: object
...
  types.RepoTag:
    properties:
      built_in:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      group:
        type: string
      name:
        type: string
      show_name:
        type: string
      updated_at:
        type: string
    type: object
  types.Repository:
    properties:
      http_clone_url:
        type: string
      ssh_clone_url:
        type: string
    type: object
  types.RepositorySource:
    enum:
    - opencsg
    - local
    - huggingface
    type: string
    x-enum-varnames:
    - OpenCSGSource
    - LocalSource
    - HuggingfaceSource
  types.RepositorySyncStatus:
    enum:
    - pending
    - inprogress
    - failed
    - completed
    type: string
    x-enum-varnames:
    - SyncStatusPending
    - SyncStatusInProgress
    - SyncStatusFailed
    - SyncStatusCompleted
  types.RepositoryType:
    enum:
    - model
    - dataset
    - space
    - code
    - prompt
    - ""
    type: string
    x-enum-varnames:
    - ModelRepo
    - DatasetRepo
    - SpaceRepo
    - CodeRepo
    - PromptRepo
    - UnknownRepo
...
  types.Response:
    properties:
      data: {}
      msg:
        type: string
    type: object
  types.ResponseWithTotal:
    properties:
      data: {}
      msg:
        type: string
      total:
        type: integer
    type: object
...
  types.Space:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      cover_image_url:
        type: string
      created_at:
        type: string
      default_branch:
        type: string
      description:
        example: ""
        type: string
      endpoint:
        description: the serving endpoint url
        example: https://localhost/spaces/myname/myspace
        type: string
      env:
        type: string
      hardware:
        type: string
      id:
        type: integer
      license:
        example: MIT
        type: string
      like_count:
        type: integer
      name:
        example: space_name_1
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        example: ""
        type: string
      path:
        example: user_or_org_name/space_name_1
        type: string
      private:
        type: boolean
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      sdk:
        description: like gradio,steamlit etc
        example: "1"
        type: string
      sdk_version:
        example: v0.1
        type: string
      secrets:
        type: string
      sku:
        type: string
      source:
        $ref: '#/definitions/types.RepositorySource'
      status:
        description: deploying, running, failed
        type: string
      svc_name:
        type: string
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      template:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
      username:
        example: creator_user_name
        type: string
    type: object
...
  types.UpdateDatasetReq:
    properties:
      description:
        type: string
      nickname:
        example: model display name
        type: string
      private:
        example: false
        type: boolean
    type: object
...
  types.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      can_change_username:
        type: boolean
      email:
        type: string
      homepage:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      nickname:
        type: string
      orgs:
        items:
          $ref: '#/definitions/types.Organization'
        type: array
      phone:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /datasets:
    get:
      consumes:
      - application/json
      description: get visiable datasets for current user
      parameters:
      - description: current user
        in: query
        name: current_user
        type: string
      - description: search text
        in: query
        name: search
        type: string
      - description: filter by task tag
        in: query
        name: task_tag
        type: string
      - description: filter by framework tag
        in: query
        name: framework_tag
        type: string
      - description: filter by license tag
        in: query
        name: license_tag
        type: string
      - description: filter by language tag
        in: query
        name: language_tag
        type: string
      - description: sort by
        in: query
        name: sort
        type: string
      - description: source
        enum:
        - opencsg
        - huggingface
        - local
        in: query
        name: source
        type: string
      - default: 20
        description: per
        in: query
        name: per
        type: integer
      - default: 1
        description: per page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Dataset'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get Visiable datasets for current user
      tags:
      - Dataset
    post:
      consumes:
      - application/json
      description: create a new dataset
      parameters:
      - description: current user, the owner
        in: query
        name: current_user
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.CreateDatasetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Dataset'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Create a new dataset
      tags:
      - Dataset
  /datasets/{namespace}/{name}:
    delete:
      consumes:
      - application/json
      description: delete a exists dataset
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current user, the owner
        in: query
        name: current_user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Delete a exists dataset
      tags:
      - Dataset
    get:
      consumes:
      - application/json
      description: get dataset detail
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Dataset'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get dataset detail
      tags:
      - Dataset
    put:
      consumes:
      - application/json
      description: update a exists dataset
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current user, the owner
        in: query
        name: current_user
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateDatasetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/database.Dataset'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Update a exists dataset
      tags:
      - Dataset
  /datasets/{namespace}/{name}/all_files:
    get:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current user
        in: query
        name: current_user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.File'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get all files of a dataset
      tags:
      - Dataset
  /datasets/{namespace}/{name}/relations:
    get:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Relations'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get dataset related assets
      tags:
      - Dataset
...
securityDefinitions:
  ApiKey:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"